relocate v2 flag block event log function.
